// src/API.ts
export type Post = {
  __typename: "Post",
  id: string,
  title: string,
  content: string,
  image?: string | null,
  createdAt: string,
  updatedAt: string,
};
// ...
export type ModelPostConnection = {
  __typename: "ModelPostConnection",
  items:  Array<Post | null >,
  nextToken?: string | null,
};
// ...
export type ListPostsQuery = {
  listPosts?:  {
    __typename: "ModelPostConnection",
    items:  Array< {
      __typename: "Post",
      id: string,
      title: string,
      content: string,
      image?: string | null,
      createdAt: string,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};